requirements.txt
websocket_client
requests==2.24.0
Flask==1.1.2
tornado==5.1.1
